app/requirements.txt
aws-cdk.core
aws-cdk.aws_iam
aws-cdk.aws_lambda
aws-cdk.aws_lambda_event_sources
aws-cdk.aws_dynamodb


#test
requests
requests_toolbelt